% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functins.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots all fatalities by location
\code{fars_map_state} will plot a map of fatalities by location for a a state
and a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number corresponding to a state}

\item{year}{a number corresponding to a year}
}
\value{
returns a map with a plot for fatalities for a state and year if no
fatalities in a given state in a year a message "no accidents to plot" will
be displayed as map
}
\description{
Plots all fatalities by location
\code{fars_map_state} will plot a map of fatalities by location for a a state
and a year
}
\note{
Errors will be thrown if the argument years are non-numeric
Or if the file does not exist in the directory
Or is the file is not a csv file
If state number is invalid
If state number is a non numeric
}
\examples{
fars_map_state(5,2015)

}
