% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functins.R
\name{fars_read}
\alias{fars_read}
\title{National High Safety Fatality Report Data Reader
\code{fars_read} will read data from the National Highway Traffic Safety
Administration's Fatality Analysis Reporting System into a dplyr
tbl_df data type}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string representation of the filename where the data
is stored and must be a type of csv}
}
\value{
code will search for the filename is th filename exists the data
will be imported as a dplyr tbl_df
}
\description{
National High Safety Fatality Report Data Reader
\code{fars_read} will read data from the National Highway Traffic Safety
Administration's Fatality Analysis Reporting System into a dplyr
tbl_df data type
}
\note{
Errors will be thrown if the argument filename is not a string
Or if the file does not exist in the directory
Or is the file is not a csv file
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
