% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functins.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Fatality report reader year by year
This function takes a vector of years as an input and goes through the years
one by one. The function takes one of the years and uses it as input to
\code{make_filename()} and creates a filename of the format
\code{accident_year.csv.bz}. Then function calls \code{fars_read()} to read in
the file as dply tbl_data then restructures it to be one year per column.
If the file is not available the function will return
\code{"invalid year: ..."}.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector representation the years of the files that you want to
load into R}
}
\value{
\code{fars_read_years} will search for all of the file years in the
inputted years vector and return the data if available if present an error
will be thrown
}
\description{
Fatality report reader year by year
This function takes a vector of years as an input and goes through the years
one by one. The function takes one of the years and uses it as input to
\code{make_filename()} and creates a filename of the format
\code{accident_year.csv.bz}. Then function calls \code{fars_read()} to read in
the file as dply tbl_data then restructures it to be one year per column.
If the file is not available the function will return
\code{"invalid year: ..."}.
}
\note{
View the \code{make_filename()} documentation for more info one how the
filenames are created.

Errors will be thrown if the argument years are non-numeric
Or if the file does not exist in the directory
Or is the file is not a csv file
}
\examples{
\dontrun{
fars_read_years(2013:2015)
fars_read_years(c(2013,2015))
}
}
