% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functins.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes information about the yearly fatalities
\code{fars_summarize_years} will provide a monthly and yearly count of
fatalities over the inputted years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of inputted years}
}
\value{
returns a tbl_df of monthly and yearly count of fatalities over the
inputted years
}
\description{
Summarizes information about the yearly fatalities
\code{fars_summarize_years} will provide a monthly and yearly count of
fatalities over the inputted years
}
\note{
Errors will be thrown if the argument years are non-numeric
Or if the file does not exist in the directory
Or is the file is not a csv file
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013,2015))
}
}
